==> 04_slide_deck.Rmd <==
---
title: \textcolor{black}{New Apprenticeship Registrations and Apprenticeship Certifications}
subtitle: \textcolor{black}{`r paste("for",format(readRDS(here::here("processed_data","max_date.rds")), format="%B %Y"))`}
author: "Labour Market Information Office"
institute: "Contact: Richard.Martin@gov.bc.ca"
date: "`r lubridate::today()`"
header-includes: 
- \usepackage{booktabs}
- \usepackage{colortbl}
- \usepackage{makecell}
- \usepackage{xcolor}
- \usepackage{longtable}
- \definecolor{bcblue}{HTML}{1f4181}
- \definecolor{bcyellow}{HTML}{ffba2a}
- \setbeamercolor*{palette primary}{use=structure,fg=white,bg=bcblue}
- \setbeamercolor*{palette quaternary}{use=structure,fg=white,bg=bcyellow}
- \setbeamercolor{title}{fg=bcblue}
- \setbeamercolor{frametitle}{fg=bcblue}
- \useoutertheme{split}
output:
  beamer_presentation:
    theme: "Boadilla"
    fonttheme: "structurebold"
urlcolor: blue
linkcolor: blue
classoption: "aspectratio=169"
always_allow_html: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE)
library(tidyverse)
library(fpp3)
library(here)
library(janitor)
library(readxl)
library(cansim)
library(kableExtra)
library(ggfortify)
library(grid)
library(tidyquant)
library(plotly)
source(here("R","functions.R"))
#constants---------------------------
alpha <- .75 # the exponential smoothing weight
#functions-----------------------
get_base <- function(tbbl, alpha){
  value <- tbbl|>
    mutate(raw=(1-alpha)^exponent,
           weight=raw/sum(raw),
           weighted_reg=new_reg*weight)|>
    summarize(sum(weighted_reg))|>
    pull()
  date <- max(tbbl$mid_point)
  list(value=value, date=date)
}

make_forecast <- function(base, agf){
  tbbl <- tibble(exponent=1:11,
                 date=base$date+years(1:11),
                 new_reg_fcast=base$value*agf^exponent)|>
    mutate(cagr=agf-1)|>
    select(-exponent)
}  
plot_fcast <- function(tbbl){
  ggplot(tbbl, aes(date, 
                                value, 
                                colour=series,
                                group=series,
                                text=paste("Date:",
                                           date,
                                           "\n Value:",
                                           scales::comma(value, accuracy = 1),
                                           "\n CAGR:",
                                           scales::percent(cagr, accuracy = .01)
                                           )
                                )
                )+
  geom_line()+
  facet_wrap(~fct_reorder(group, value, .desc=TRUE), scales="free_y", nrow=4)+
  labs(colour=NULL,
       x=NULL,
       y=NULL)+
  scale_y_continuous(labels=scales::comma)
}

#read in data-------------------------
reg_and_employment <- read_rds(here("processed_data","reg_and_employment.rds"))
regions <- unique(reg_and_employment$drname)

theme_set(theme_grey(base_size = 15))
mapping <- read_csv(here("current_data","mapping","mapping.csv"))
new_reg <- read_clean_join("New")

by_year_and_age <- new_reg%>%
  group_by(year, age_group_registration_start)%>%
  summarize(count=sum(count, na.rm=TRUE))%>%
  filter(!stringr::str_detect(age_group_registration_start, "="),
         age_group_registration_start!="NULL")

pop <- read_rds(here::here("current_data","lmo","pop.rds"))

prop_reg <- full_join(by_year_and_age, pop, by=c("year"="year","age_group_registration_start"="age_group"))%>%
  mutate(prop=count/value)

new_reg_total <- new_reg%>%
  just_aggregate()%>%
  fix_dates()
max_date <- yearmonth(max(new_reg_total$date))
max_year <- year(max_date)
stc_new_reg <- new_reg%>%
  filter_and_aggregate(stc, "equal", "STC Trades")%>%
  fix_dates()
cofq <- read_clean_join("CofQ")
cofq_total <- cofq%>%
  just_aggregate(program_type)%>%
  fix_dates()
cofq_app <- cofq_total%>%
  filter(program_type=="Apprenticeship")
cofq_chal <- cofq_total%>%
  filter(program_type=="Challenge")

#make the forecasts-------------

reg_emp_alpha <- reg_and_employment|>
  mutate(base=map(backward_annualized, get_base, alpha),
         forecast=map2(base, agf, make_forecast)
         )
historic <- reg_emp_alpha |>
  select(drname, group, backward_annualized)|>
  unnest(backward_annualized)|>
  select(drname, group, date=mid_point, value=new_reg)|>
  mutate(series="Backward Annualized Registrations")

forecast <- reg_emp_alpha|>
  select(drname, group, forecast)|>
  unnest(forecast)|>
  select(drname, group, date, value=new_reg_fcast, cagr)|>
  mutate(series="Forecast")

all_data <- bind_rows(historic, forecast)|>
    arrange(drname, group, date)
```

## Highlights

- Comparisons to same month last year:
  - New apprenticeship registrations in `r max_date` were `r describe_change(get_growth(new_reg_total, years, 1))`.
  - Registrations in the Skilled Trades Certification (STC) trades occupations in `r max_date` were `r describe_change(get_growth(stc_new_reg, years, 1))`.
  - Certificates of Qualification issued in `r max_date` were `r describe_change(get_growth(cofq_app, years, 1))` for Apprentices.
  - Certificates of Qualification issued in `r max_date` were `r describe_change(get_growth(cofq_chal, years, 1))` for Trade Qualifiers.

- Forecast for next year:
  - Relative to last year, new apprenticeship registrations in the next 12 months are forecast to be `r scales::percent(head(forecast$cagr[forecast$drname=="British Columbia" & forecast$group=="All groups"], n=1), accuracy=.01)` higher.

## New Apprentice Registrations: [Total*](#Total)
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{Put some commentary here} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
decomp_plot(new_reg_total)
```
::::
::::  {.column width=.35}
- Compared to the same month last year, New apprenticeship registrations were `r describe_change(get_growth(new_reg_total, years, 1))`.
- Compared to last month, New apprenticeship registrations were `r describe_change(get_growth(new_reg_total, months, 1))`.
::::
:::

## New Apprentice Registrations: [Skilled Trades Certification*](#STC) 
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{Put some commentary here} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
  decomp_plot(stc_new_reg)
```
::::
::::  {.column width=.35}
- Compared to the same month last year, STC new registrations were `r describe_change(get_growth(stc_new_reg, years, 1))`.
- Compared to last month, STC new registrations were `r describe_change(get_growth(stc_new_reg, months, 1))`.
::::
:::

## Certificate of Qualification: Total
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{Put some commentary here} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
cofq_agg <- cofq_total%>%
  group_by(date)%>%
  summarize(count=sum(count))
decomp_plot(cofq_agg)
```
::::
::::  {.column width=.35}
- Compared to the same month last year, Certificates of Qualification were `r describe_change(get_growth(cofq_agg, years, 1))`.
- Compared to last month,  Certificates of Qualification were `r describe_change(get_growth(cofq_agg, months, 1))`.
::::
:::

## Certificate of Qualification: [Apprenticeship route*](#A1)
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{Put some commentary here} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
decomp_plot(cofq_app)
```
::::
::::  {.column width=.35}
- Compared to the same month last year, Certificates of Qualification (apprentice route) were `r describe_change(get_growth(cofq_app, years, 1))`.
- Compared to last month,  Certificates of Qualification (apprentice route) were `r describe_change(get_growth(cofq_app, months, 1))`.
::::
:::

## Certificate of Qualification: [Challenge route*](#A1)
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{Put some commentary here} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
decomp_plot(cofq_chal)
```
::::
::::  {.column width=.35}
- Compared to the same month last year, Certificates of Qualification (challenge route) were `r describe_change(get_growth(cofq_chal, years, 1))`.
- Compared to last month, Certificates of Qualification (challenge route) were `r describe_change(get_growth(cofq_chal, months, 1))`.
::::
:::

## New Apprenticeship Registrations: [Total*](#Total)
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{New Forecast Method} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
all_groups <- all_data|>
  filter(drname=="British Columbia",
         group=="All groups")

all_groups_fyof <- all_groups|>
  filter(series=="Forecast")|>
  filter(date==min(date))|>
  pull(value)|>
  scales::comma()

plot_fcast(all_groups)
```
::::
::::  {.column width=.35}
- Base value is an exponential weighted average of historic new registrations.
- New registrations assumed to grow at same rate as associated occupations (from LMO): `r scales::percent(tail(all_groups$cagr, n=1), accuracy=.01)`
- Over the next 12 months new registrations are forecast to be `r all_groups_fyof`.
::::
:::


## New Apprenticeship Registrations: [Skilled Trades Certification*](#STC)
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{New Forecast Method} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
stc_bc <- all_data|>
  filter(drname=="British Columbia",
         group=="STC Trades")

stc_bc_fyof <- stc_bc|>
  filter(series=="Forecast")|>
  filter(date==min(date))|>
  pull(value)|>
  scales::comma()

  plot_fcast(stc_bc)
```
::::
::::  {.column width=.35}
- Base value is an exponential weighted average of historic new registrations.
- New registrations assumed to grow at same rate as associated occupations (from LMO): `r scales::percent(tail(stc_bc$cagr, n=1), accuracy=.01)`
- Over the next 12 months new registrations are forecast to be `r stc_bc_fyof`.
::::
:::

## New Apprenticeship Registrations: [Structural Building Trades*](#SBT)
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{New Forecast Method} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
sbt_bc <- all_data|>
  filter(drname=="British Columbia",
         group=="Structural Building Trades")

sbt_bc_fyof <- sbt_bc|>
  filter(series=="Forecast")|>
  filter(date==min(date))|>
  pull(value)|>
  scales::comma()

plot_fcast(sbt_bc)
```
::::
::::  {.column width=.35}
- Base value is an exponential weighted average of historic new registrations.
- New registrations assumed to grow at same rate as associated occupations (from LMO): `r scales::percent(tail(sbt_bc$cagr, n=1), accuracy=.01)`
- Over the next 12 months new registrations are forecast to be `r sbt_bc_fyof`.
::::
:::


## Population Comparison
::: columns 
:::: {.column width=.05}
::::
:::: {.column width=.6}
### \textcolor{bcblue}{Ratio of New Registrations to Population} 
```{r, echo=FALSE, fig.show='hold', fig.height=6}
  ggplot(prop_reg, aes(age_group_registration_start, prop, fill=as.factor(year)))+
  geom_col(position = "dodge")+
  scale_fill_viridis_d()+
  labs(x="Age Group", y="",fill="", caption="Population data from Statistics Canada table 17-10-0005-01")+
  scale_y_continuous(labels=scales::percent)
```
::::
::::  {.column width=.35}
- People under 30 are those most likely to register for an apprenticeship.
- People of all ages were less likely to register for an apprenticeship in 2020. 
- For all age groups the proportion of new registrations is no lower than pre-COVID. 
::::
:::

## Appendix 1 {#A1}

- Source Code: [Github](https://github.com/bcgov/Industry_Trades_Authority)
- New registration forecasts are the product of the historic ratio between new registrations and employment and the forecast of trades employment from the last published edition of the BC Labour Market Outlook (LMO). 
- The current year forecast (`r year(max(new_reg_total$date))`) is a weighted average of new registrations thus far and the demand driven forecast described above.
- Historical data on New Apprenticeship Registrations and Certificates of Qualification was provided by SkilledTradesBC.
- Certificates of Qualification can be earned through two pathways: The completion of an apprenticeship program (Apprentice route, called “Apprentices”) or by challenging the certification exam (Challenge route, called “Trade Qualifiers”)

## Appendix 2: STC trades {#STC}

- These are the first 10 trades to undergo mandatory certification.

```{r}
stc <- mapping%>%
  filter(stc=="STC Trades")%>%
  select(noc_code, description)
colnames(stc) <- c("NOC Code", "Description")
kable(stc, booktabs = T)%>%
  kable_styling(latex_options = "striped")
```

## Appendix 3: Functional Trade Groups: {#FTG}

```{r}
ftg <- new_reg%>%
  select(functional_trades_group)%>%
  distinct()%>%
  na.omit()
colnames(ftg) <- make_title(colnames(ftg))
kable(ftg, booktabs = T)%>%
  kable_styling(latex_options = "striped")
```

## Appendix 4: Structural Building Trades: {#SBT}

```{r}
sbt <- new_reg%>%
  filter(functional_trades_group=="Structural Building Trades")%>%
  select(noc_code, description)%>%
  distinct()%>%
  na.omit()
colnames(sbt) <- c("NOC Code", "Description")
kable(sbt, booktabs = T)%>%
  kable_styling(latex_options = "striped", font_size = 7)
```

## Appendix 5: All trade occupations {#Total}

```{r, message=FALSE}
all_trades <- new_reg%>%
  filter(trades=="Trades")%>%
  select(noc_code, description)%>%
  distinct()

first <- head(all_trades, n=ceiling(nrow(all_trades)/2))
second <- tail(all_trades, n=ceiling(nrow(all_trades)/2))
all <- bind_cols(first, second)

kable(all, 
      booktabs = T,
      col.names = c("NOC Code",
                           "Description",
                           "NOC Code",
                           "Description"))%>%
  kable_styling(latex_options = "striped", font_size = 5)
```
